Call recording encoder/format choice 2/3
Adds a user option between NB AMR (the default) or HE-AAC for the encoding of call recording audio. Also changes the output format of the file appropriately (NB AMR now uses the NB_AMR output format and HE_AAC chooses the MPEG_4 output. File extensions are adjust accordingly as well).
This was in reponse to people complaining the quality of call recording audio is poor. It's noticibly better using HE-AAC, but filesizes are larger.

PS: Imported from kitkat and made working for lollipop.

Change-Id: Ied0e361e8f4691a3a81b3385ffbc30618f658ac1
Signed-off-by: BlackDragon <blackdragon.fusionteam@gmail.com>